refactor(mcp): split the tool catalog by operations tier
toolCatalog was a single 860-line function inside a 1764-line file that
also held the writer interfaces, the icon map, registration, the handlers,
the result types and the argument decoding — six concerns in the one file
of the package that otherwise separates them cleanly.

The catalog moves out into tools_reads.go, tools_operational.go,
tools_configuration.go and tools_impactful.go, concatenated by
toolCatalog. Tier is the axis because it is what registerTools filters on,
so a tool's file now says what gates it.

The per-resource check*Write/check*Read helpers move to acl.go, beside
checkRead and the filters they belong with. tools.go keeps the interfaces,
registration, handlers and result types, and drops to 794 lines.

No behaviour change: same tools, same tiers, same handlers.
