+++ internal/mcp/tools.go
@@ -5,6 +5,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"maps"
+	"slices"
 	"strings"
 
 	"github.com/docker/docker/api/types/container"
@@ -12,7 +13,6 @@ import (
 	"github.com/docker/docker/api/types/swarm"
 	mcplib "github.com/mark3labs/mcp-go/mcp"
 
-	"github.com/radiergummi/cetacean/internal/auth"
 	"github.com/radiergummi/cetacean/internal/cluster"
 	"github.com/radiergummi/cetacean/internal/config"
 	"github.com/radiergummi/cetacean/internal/docker"
@@ -279,881 +279,12 @@ func (s *Server) registerTools() {
 // mcp-go's NewTool defaults destructive and open-world to true, which is the
 // wrong shape for a closed cluster-management surface.
 func (s *Server) toolCatalog() []toolDef {
-	return []toolDef{
-		// Tier 0 — parameterized reads, no gating.
-		{
-			tool: mcplib.NewTool(
-				"get_logs",
-				mcplib.WithToolTitle("Read service or task logs"),
-				mcplib.WithDescription(
-					"Fetch recent log lines from one service, one task, a whole stack, or the whole cluster. Name exactly one scope: `service` merges the output of every live replica; `task` reads a single replica and is the only way to reach one that has already exited; `stack` and `cluster` merge across services server-side, so grepping the cluster is one call rather than one per service, and every line says which service it came from. Swarm discards a task's output once its record falls out of the history window (five per replica slot by default), so on a service restarting in a loop that history is only seconds deep and should be read first. Narrow with `contains` for a substring, `level` for a minimum severity and `since` for a start time — on a wide read `contains` is what keeps the answer small. A wide read covers at most 25 services and reports any it could not reach in `errors` rather than failing. This is a one-shot read; for live tails subscribe to the cetacean://services/{id}/logs resource.",
-				),
-				mcplib.WithOutputSchema[LogResourceResponse](),
-				mcplib.WithReadOnlyHintAnnotation(true),
-				mcplib.WithDestructiveHintAnnotation(false),
-				mcplib.WithIdempotentHintAnnotation(true),
-				mcplib.WithOpenWorldHintAnnotation(false),
-				mcplib.WithString("service",
-					mcplib.Description(
-						"Service ID or name to read logs from, merged across its live replicas. Mutually exclusive with `task`.",
-					),
-				),
-				mcplib.WithString("task",
-					mcplib.Description(
-						"Task ID to read logs from, including a task that has already exited. Mutually exclusive with `service`.",
-					),
-				),
-				mcplib.WithString("stack",
-					mcplib.Description(
-						"Stack name; merges the logs of every service in it. Mutually exclusive with `service` and `task`.",
-					),
-				),
-				mcplib.WithBoolean("cluster",
-					mcplib.Description(
-						"Read every service in the cluster, merged. Mutually exclusive with `service`, `task` and `stack`.",
-					),
-				),
-				mcplib.WithString("contains",
-					mcplib.Description(
-						"Keep only lines containing this substring (case-insensitive). Applied server-side, so a wide read costs the matches rather than every line.",
-					),
-				),
-				mcplib.WithNumber(
-					"tail",
-					mcplib.Description(
-						"Number of recent log lines to return per replica. Default 100.",
-					),
-				),
-				mcplib.WithString(
-					"since",
-					mcplib.Description(
-						"RFC 3339 timestamp; only return lines emitted strictly after this time.",
-					),
-				),
-				mcplib.WithString(
-					"level",
-					mcplib.Description(
-						"Minimum log level (debug, info, warn, error). Best-effort — depends on the service emitting structured levels.",
-					),
-				),
-			),
-			tier:    config.OpsReadOnly,
-			handler: s.toolGetLogs,
-			widget:  "logs",
-		},
-		{
-			tool: mcplib.NewTool(
-				"find",
-				mcplib.WithToolTitle("Find cluster resources"),
-				mcplib.WithDescription(
-					"Locate cluster resources. Give `type` (nodes, services, tasks, stacks, configs, secrets, networks, or volumes) to enumerate that type, paged, optionally narrowed by query/state/stack/node/image/label; the returned records are the same ones the cetacean:// resources expose, filtered to what the caller may see. Omit `type` and give `query` to search by name, label, or image reference across every type at once, returned as one list sorted by name, each row carrying its own type; there `limit` caps the matches per type and there is no paging, so `total` counts every match across the cluster and `counts` breaks it down per type — narrow with `type` to page through one of them. Use the cross-type search to locate a resource before fetching its details or applying a write; use a typed listing to browse or tabulate a whole type.",
-				),
-				mcplib.WithOutputSchema[findResult](),
-				mcplib.WithReadOnlyHintAnnotation(true),
-				mcplib.WithDestructiveHintAnnotation(false),
-				mcplib.WithIdempotentHintAnnotation(true),
-				mcplib.WithOpenWorldHintAnnotation(false),
-				mcplib.WithString(
-					"type",
-					mcplib.Description(
-						"Resource type to enumerate: nodes, services, tasks, stacks, configs, secrets, networks, or volumes. Omit to search across every type at once (requires `query`).",
-					),
-				),
-				mcplib.WithString(
-					"query",
-					mcplib.Description(
-						"Substring to match against resource names — and, when `type` is omitted, also labels and image references. Case-insensitive. Required when `type` is omitted.",
-					),
-				),
-				mcplib.WithString("state",
-					mcplib.Description(
-						"With `type`, keep only rows whose derived state matches exactly (case-insensitive).",
-					),
-				),
-				mcplib.WithString("stack",
-					mcplib.Description(
-						"With `type`, keep only rows in this stack namespace (exact match, case-insensitive).",
-					),
-				),
-				mcplib.WithString("node",
-					mcplib.Description(
-						"With `type: tasks`, keep only rows running on a node whose hostname contains this (case-insensitive).",
-					),
-				),
-				mcplib.WithString("image",
-					mcplib.Description(
-						"With `type: services`, keep only rows whose image contains this (case-insensitive).",
-					),
-				),
-				mcplib.WithString("label",
-					mcplib.Description(
-						"With `type`, keep only rows carrying this label: `key` for presence, `key=value` for an exact value.",
-					),
-				),
-				mcplib.WithNumber("limit",
-					mcplib.Description(
-						"With `type`, the maximum records to return (default and maximum 200). Without `type`, the maximum matches per resource type in the cross-type search (default 3).",
-					),
-				),
-				mcplib.WithNumber("offset",
-					mcplib.Description(
-						"With `type`, records to skip, for paging. Default 0. Ignored in the cross-type search.",
-					),
-				),
-				mcplib.WithBoolean("raw",
-					mcplib.Description(
-						"With `type`, add each match's untouched resource record to the result, under `raw`, beside the compact rows. Default false.",
-					),
-				),
-			),
-			tier:    config.OpsReadOnly,
-			handler: s.toolFind,
-			widget:  "table",
-		},
-		{
-			tool: mcplib.NewTool(
-				"describe",
-				mcplib.WithToolTitle("Describe one cluster resource"),
-				mcplib.WithDescription(
-					"Return everything needed to act on one resource: its derived state, the reason Swarm gave for that state when it is not healthy, how long the state has held, the type-specific facts (a service's image, replica counts, reserved CPU in cores and memory in bytes, ports, placement constraints and environment variable *names*; a node's role, availability, capacity; a network's driver and subnets), the resources it references or is referenced by, and the recent task failures behind a failing state. Name the type in the singular (service, node, task, stack, config, secret, network, volume) and give an ID or a name. Use this instead of a resource read when following up on a finding from find or get_recommendations; secret payloads and environment variable values are never returned.",
-				),
-				mcplib.WithOutputSchema[describeResult](),
-				mcplib.WithReadOnlyHintAnnotation(true),
-				mcplib.WithDestructiveHintAnnotation(false),
-				mcplib.WithIdempotentHintAnnotation(true),
-				mcplib.WithOpenWorldHintAnnotation(false),
-				mcplib.WithString("type",
-					mcplib.Required(),
-					mcplib.Description(
-						"Resource type, singular: service, node, task, stack, config, secret, network or volume.",
-					),
-				),
-				mcplib.WithString("id",
-					mcplib.Required(),
-					mcplib.Description(
-						"The resource's ID, or its name (hostname for a node, name for a stack or volume).",
-					),
-				),
-				mcplib.WithBoolean("raw",
-					mcplib.Description(
-						"Add the untouched Docker record to the result, under `raw`, beside the digest. Default false.",
-					),
-				),
-			),
-			tier:    config.OpsReadOnly,
-			handler: s.toolDescribe,
-		},
-		{
-			tool: mcplib.NewTool(
-				"get_topology",
-				mcplib.WithToolTitle("Get cluster topology"),
-				mcplib.WithDescription(
-					"Return the cluster as a graph. The network view joins services to the overlay networks they attach to, answering which services can reach each other; the placement view joins cluster nodes to the services they run, answering what runs where; the drain-impact view takes a `node` and joins the services running on it to the nodes that could take them, answering what would move if you drained it and what could not. In drain-impact a service with no edges is stranded and its `detail` names the placement constraint that blocked it — check this before draining a node rather than discovering it afterwards. Use this instead of listing services and nodes separately when the question is about relationships.",
-				),
-				mcplib.WithOutputSchema[cluster.TopologyGraph](),
-				mcplib.WithReadOnlyHintAnnotation(true),
-				mcplib.WithDestructiveHintAnnotation(false),
-				mcplib.WithIdempotentHintAnnotation(true),
-				mcplib.WithOpenWorldHintAnnotation(false),
-				mcplib.WithString(
-					"view",
-					mcplib.Description(
-						"Which projection to return: \"network\" (services joined to overlay networks, the default), \"placement\" (cluster nodes joined to the services they run), or \"drain-impact\" (the services on one node joined to the nodes that could take them; requires `node`).",
-					),
-				),
-				mcplib.WithString(
-					"node",
-					mcplib.Description(
-						"Node ID or hostname to assess. Required by the \"drain-impact\" view and ignored by the others.",
-					),
-				),
-			),
-			tier:    config.OpsReadOnly,
-			handler: s.toolGetTopology,
-			widget:  "topology",
-		},
-		{
-			tool: mcplib.NewTool(
-				"get_metrics",
-				mcplib.WithToolTitle("Chart a resource metric"),
-				mcplib.WithDescription(
-					"Return a time series of CPU, memory or network use, over the last hour, six hours, day or week. Chart one resource by naming `target` \"service\" or \"node\" with its `id`. Rank instead by asking for `top`: `target` \"cluster\" ranks the busiest services cluster-wide (or nodes, with `by`), and `target` \"node\" with an `id` and `top` ranks the services running on that one host — which is how to answer what is using the most CPU, and why a given node is hot. A ranking returns one series per member, named the way the cluster names it. Requires Prometheus (and cAdvisor for service metrics, node-exporter for node metrics); a Cetacean without them reports that metrics are unavailable rather than returning empty series. Cetacean owns the queries — name a target and a metric, not PromQL.",
-				),
-				mcplib.WithOutputSchema[metricsResult](),
-				mcplib.WithReadOnlyHintAnnotation(true),
-				mcplib.WithDestructiveHintAnnotation(false),
-				mcplib.WithIdempotentHintAnnotation(true),
-				mcplib.WithOpenWorldHintAnnotation(false),
-				mcplib.WithString("target",
-					mcplib.Required(),
-					mcplib.Description(
-						"What to measure: \"service\", \"node\", or \"cluster\" to rank across the whole cluster.",
-					),
-				),
-				mcplib.WithString("id",
-					mcplib.Description(
-						"The service (ID or name) or node (ID or hostname) to measure. Required for the \"service\" and \"node\" targets; the \"cluster\" target takes none.",
-					),
-				),
-				mcplib.WithNumber("top",
-					mcplib.Description(
-						"Return a ranking of the busiest members rather than one resource's own series: the top N (default 5, maximum 10). With target \"cluster\" it ranks the whole cluster; with target \"node\" it ranks the services running on that node.",
-					),
-				),
-				mcplib.WithString("by",
-					mcplib.Description(
-						"What a cluster-wide ranking ranks: \"service\" (default) or \"node\". Ignored by every other target.",
-					),
-				),
-				mcplib.WithString("metric",
-					mcplib.Description(
-						"Which metric: \"cpu\" (default), \"memory\" or \"network\". Service CPU is percent of a core and service memory is bytes; node CPU and memory are percentages. Network is two series, receive and transmit, in bytes per second — a ranking sums the two, since it asks which member moves the most traffic.",
-					),
-				),
-				mcplib.WithString("range",
-					mcplib.Description(
-						"Window to chart: \"1h\" (default), \"6h\", \"24h\" or \"7d\".",
-					),
-				),
-			),
-			tier:    config.OpsReadOnly,
-			handler: s.toolGetMetrics,
-			widget:  "metrics",
-		},
-		{
-			tool: mcplib.NewTool(
-				"get_recommendations",
-				mcplib.WithToolTitle("Get cluster recommendations"),
-				mcplib.WithDescription(
-					"Return what Cetacean's recommendation engine currently finds: over- and under-provisioned resources, missing health checks or restart policies, flaky services, single-replica risks, manager workload imbalance and uneven node distribution. Each entry carries a severity, the resource it concerns and why it was raised. Only findings about resources the caller may read are returned.",
-				),
-				mcplib.WithOutputSchema[recommendationsResult](),
-				mcplib.WithReadOnlyHintAnnotation(true),
-				mcplib.WithDestructiveHintAnnotation(false),
-				mcplib.WithIdempotentHintAnnotation(true),
-				mcplib.WithOpenWorldHintAnnotation(false),
-				mcplib.WithString("severity",
-					mcplib.Description(
-						"Return only findings of this severity: \"critical\", \"warning\" or \"info\". Omit for all of them.",
-					),
-				),
-			),
-			tier:    config.OpsReadOnly,
-			handler: s.toolGetRecommendations,
-			widget:  "recommendations",
-		},
-		{
-			tool: mcplib.NewTool(
-				"get_events",
-				mcplib.WithToolTitle("Read the cluster change timeline"),
-				mcplib.WithDescription(
-					"Return what changed in the cluster and when — resource creates, updates and deletes — newest first, narrowed by time, type or a single resource. Use it to answer what happened overnight, whether a fault is new, and what else changed around the moment something broke; pair it with get_logs, which returns the same entry shape, to read changes and output on one timeline. Filtering by `types` is usually necessary: a service restarting in a loop produces a task event every few seconds and buries everything else.",
-				),
-				mcplib.WithOutputSchema[eventsResult](),
-				mcplib.WithReadOnlyHintAnnotation(true),
-				mcplib.WithDestructiveHintAnnotation(false),
-				mcplib.WithIdempotentHintAnnotation(true),
-				mcplib.WithOpenWorldHintAnnotation(false),
-				mcplib.WithString("since",
-					mcplib.Description(
-						"RFC 3339 timestamp; only entries strictly after this time.",
-					),
-				),
-				mcplib.WithString("until",
-					mcplib.Description(
-						"RFC 3339 timestamp; only entries at or before this time.",
-					),
-				),
-				mcplib.WithArray("types",
-					mcplib.Description(
-						"Resource types to include: node, service, task, stack, config, secret, network, volume. Omit for all.",
-					),
-					mcplib.Items(map[string]any{"type": "string"}),
-				),
-				mcplib.WithString("resource",
-					mcplib.Description("Only entries for this resource ID."),
-				),
-				mcplib.WithNumber("limit",
-					mcplib.Description(
-						"Maximum entries to return (default 100, maximum 500).",
-					),
-				),
-			),
-			tier:    config.OpsReadOnly,
-			handler: s.toolGetEvents,
-		},
-		{
-			tool: mcplib.NewTool(
-				"get_cluster_status",
-				mcplib.WithToolTitle("Check overall cluster health"),
-				mcplib.WithDescription(
-					"Answer whether the cluster is healthy and, when it is not, name what is wrong: the services not in their desired state, the nodes that are down or draining, the rollouts in flight, and how much CPU and memory is reserved against what the cluster has. Start here — it is the one call that says whether to look further and where, and every entry carries the id and name to describe next.",
-				),
-				mcplib.WithOutputSchema[cluster.ClusterStatus](),
-				mcplib.WithReadOnlyHintAnnotation(true),
-				mcplib.WithDestructiveHintAnnotation(false),
-				mcplib.WithIdempotentHintAnnotation(true),
-				mcplib.WithOpenWorldHintAnnotation(false),
-			),
-			tier:    config.OpsReadOnly,
-			handler: s.toolGetClusterStatus,
-		},
-		{
-			tool: mcplib.NewTool(
-				"watch",
-				mcplib.WithToolTitle("Wait for a service to settle"),
-				mcplib.WithDescription(
-					"Block until a service has reached its desired state — every replica running and no rolling update in flight — then report whether it settled and how long it took. Use it after a deploy or a scale instead of describing the service in a loop. On a timeout it reports how far the rollout actually got (\"waiting: 1/3 replicas running\"), which is the answer when a deploy is stuck. The wait cannot be cancelled once started, so `timeout` is the real bound; it defaults to 60 seconds and is capped at 5 minutes.",
-				),
-				mcplib.WithOutputSchema[watchResult](),
-				mcplib.WithReadOnlyHintAnnotation(true),
-				mcplib.WithDestructiveHintAnnotation(false),
-				mcplib.WithIdempotentHintAnnotation(true),
-				mcplib.WithOpenWorldHintAnnotation(false),
-				mcplib.WithString("service",
-					mcplib.Required(),
-					mcplib.Description("Service ID or name to wait on."),
-				),
-				mcplib.WithNumber("timeout",
-					mcplib.Description(
-						"Seconds to wait before giving up (default 60, maximum 300).",
-					),
-				),
-			),
-			tier:    config.OpsReadOnly,
-			handler: s.toolWatch,
-		},
-
-		// Tier 1 — Operational.
-		{
-			tool: mcplib.NewTool(
-				"scale_service",
-				mcplib.WithToolTitle("Scale service replicas"),
-				mcplib.WithDescription(
-					"Set the desired replica count of a replicated service. Swarm reconciles asynchronously; calling with the current count is a no-op. Setting to 0 stops every task without removing the service. Returns a summary of where the service ended up: id, name, image, mode, desired replicas, running count, derived state and version.",
-				),
-				mcplib.WithReadOnlyHintAnnotation(false),
-				mcplib.WithDestructiveHintAnnotation(false),
-				mcplib.WithIdempotentHintAnnotation(true),
-				mcplib.WithOpenWorldHintAnnotation(false),
-				taskSupportOptional(),
-				mcplib.WithOutputSchema[serviceMutationResult](),
-				mcplib.WithString("id",
-					mcplib.Required(),
-					mcplib.Description("Service ID or name."),
-				),
-				mcplib.WithNumber("replicas",
-					mcplib.Required(),
-					mcplib.Description("Target replica count. Must be >= 0."),
-				),
-			),
-			tier:    config.OpsOperational,
-			handler: s.toolScaleService,
-		},
-		{
-			tool: mcplib.NewTool(
-				"update_service_image",
-				mcplib.WithToolTitle("Update service image"),
-				mcplib.WithDescription(
-					"Set the container image for a service, triggering a rolling deploy per the service's update policy. The previous spec is retained on the service history and reachable via rollback_service. Returns a summary of where the service ended up: id, name, image, mode, desired replicas, running count, derived state and version.",
-				),
-				mcplib.WithReadOnlyHintAnnotation(false),
-				mcplib.WithDestructiveHintAnnotation(true),
-				mcplib.WithIdempotentHintAnnotation(true),
-				mcplib.WithOpenWorldHintAnnotation(false),
-				taskSupportOptional(),
-				mcplib.WithOutputSchema[serviceMutationResult](),
-				mcplib.WithString("id",
-					mcplib.Required(),
-					mcplib.Description("Service ID or name."),
-				),
-				mcplib.WithString(
-					"image",
-					mcplib.Required(),
-					mcplib.Description(
-						"Container image reference including tag or digest (e.g. \"nginx:1.27\" or \"nginx@sha256:...\"). Must not be empty.",
-					),
-				),
-			),
-			tier:    config.OpsOperational,
-			handler: s.toolUpdateServiceImage,
-		},
-		{
-			tool: mcplib.NewTool("rollback_service",
-				mcplib.WithToolTitle("Roll back service to previous spec"),
-				mcplib.WithDescription(
-					"Revert a service to its previous spec, discarding the current one. Only the immediately previous spec is retained — calling twice does not unwind further. Returns a summary of where the service ended up: id, name, image, mode, desired replicas, running count, derived state and version.",
-				),
-				mcplib.WithReadOnlyHintAnnotation(false),
-				mcplib.WithDestructiveHintAnnotation(true),
-				mcplib.WithIdempotentHintAnnotation(false),
-				mcplib.WithOpenWorldHintAnnotation(false),
-				taskSupportOptional(),
-				mcplib.WithOutputSchema[serviceMutationResult](),
-				mcplib.WithString("id",
-					mcplib.Required(),
-					mcplib.Description("Service ID or name."),
-				),
-			),
-			tier:    config.OpsOperational,
-			handler: s.toolRollbackService,
-		},
-		{
-			tool: mcplib.NewTool("restart_service",
-				mcplib.WithToolTitle("Force-restart service"),
-				mcplib.WithDescription(
-					"Force a rolling restart of every task in the service by bumping its spec ForceUpdate counter. No other spec fields change. Honours the service's update policy (parallelism, delay). Returns a summary of where the service ended up: id, name, image, mode, desired replicas, running count, derived state and version.",
-				),
-				mcplib.WithReadOnlyHintAnnotation(false),
-				mcplib.WithDestructiveHintAnnotation(true),
-				mcplib.WithIdempotentHintAnnotation(false),
-				mcplib.WithOpenWorldHintAnnotation(false),
-				taskSupportOptional(),
-				mcplib.WithOutputSchema[serviceMutationResult](),
-				mcplib.WithString("id",
-					mcplib.Required(),
-					mcplib.Description("Service ID or name."),
-				),
-			),
-			tier:    config.OpsOperational,
-			handler: s.toolRestartService,
-		},
-		{
-			tool: mcplib.NewTool("remove_task",
-				mcplib.WithToolTitle("Remove (reschedule) task"),
-				mcplib.WithOutputSchema[removalResult](),
-				mcplib.WithDescription(
-					"Delete a task by ID. Swarm immediately reschedules a replacement on the parent service's behalf, so this is a forced reschedule rather than a permanent removal. Use this to evict a misbehaving task without scaling the whole service.",
-				),
-				mcplib.WithReadOnlyHintAnnotation(false),
-				mcplib.WithDestructiveHintAnnotation(true),
-				mcplib.WithIdempotentHintAnnotation(false),
-				mcplib.WithOpenWorldHintAnnotation(false),
-				mcplib.WithString("id",
-					mcplib.Required(),
-					mcplib.Description("Task ID (long-form Docker ID)."),
-				),
-			),
-			tier: config.OpsOperational,
-			handler: s.removeHandler("id", s.checkTaskWrite,
-				func(wc DockerWriteClient, ctx context.Context, id string) error {
-					return wc.RemoveTask(ctx, id)
-				}),
-		},
-
-		// Tier 2 — Configuration.
-		{
-			tool: mcplib.NewTool(
-				"update_service",
-				mcplib.WithToolTitle("Update a service's specification"),
-				mcplib.WithDescription(
-					"Change one section of a service's specification, named by `section`. `env` and `labels` take a JSON Merge Patch (RFC 7396) object — a string sets or replaces a key, null deletes it, an omitted key is preserved. The other six replace their section wholesale: pass the complete object, because a field you omit is cleared. `resources` takes ResourceRequirements (CPU/memory limits and reservations); `placement` takes Placement (constraints, preferences, max replicas per node, platforms) and may reschedule tasks onto other nodes to satisfy new constraints; `ports` takes an array of PortConfig and drops connections to any port it removes or remaps; `update-policy` and `rollback-policy` take an UpdateConfig (parallelism, delay, failure action, monitor window, max failure ratio, order) and change nothing by themselves, applying to the next spec change and to rollback_service respectively; `log-driver` takes a Driver (name plus options) and routes subsequent log lines through it. `healthcheck` takes a probe: `test` is the command as an array, for instance [`CMD`, `curl`, `-f`, `http://localhost/`], and `interval`, `timeout` and `startPeriod` are duration strings such as `10s` or `1m30s` rather than numbers; an empty `test` removes the healthcheck. `command` takes `command` (the entrypoint) and `args` (what follows it), the split Docker itself makes, and omitting one clears it. Every section except labels and the two policies triggers a rolling deploy. Returns the section as it now stands, not the whole service — describe the service for the rest.",
-				),
-				mcplib.WithOutputSchema[serviceUpdateResult](),
-				mcplib.WithReadOnlyHintAnnotation(false),
-				// Destructive because two of the sections are: replacing the
-				// ports drops the connections to a port it removes, and
-				// replacing the placement can evict running tasks. One
-				// annotation covers every section, so it has to describe the
-				// worst of them — a host deciding whether to confirm must not
-				// be told a port remap is safe because a label edit is.
-				mcplib.WithDestructiveHintAnnotation(true),
-				mcplib.WithIdempotentHintAnnotation(true),
-				mcplib.WithOpenWorldHintAnnotation(false),
-				mcplib.WithString("id",
-					mcplib.Required(),
-					mcplib.Description("Service ID or name."),
-				),
-				mcplib.WithString("section",
-					mcplib.Required(),
-					mcplib.Enum(
-						sectionEnv,
-						sectionLabels,
-						sectionResources,
-						sectionPlacement,
-						sectionPorts,
-						sectionUpdatePolicy,
-						sectionRollbackPolicy,
-						sectionLogDriver,
-						sectionHealthcheck,
-						sectionCommand,
-					),
-					mcplib.Description("Which part of the specification to change."),
-				),
-				mcplib.WithAny("value",
-					mcplib.Required(),
-					mcplib.Description(
-						"The section's new value: a merge-patch object for `env` and `labels`, an array of PortConfig for `ports`, and the replacement object for the rest. See the tool description for the shape each section expects.",
-					),
-				),
-			),
-			tier:    config.OpsConfiguration,
-			handler: s.toolUpdateService,
-		},
-		{
-			tool: mcplib.NewTool("update_node_labels",
-				mcplib.WithToolTitle("Update node labels"),
-				mcplib.WithDescription(
-					"Patch the labels on a node using JSON Merge Patch (RFC 7396) semantics: string values set or replace a key, null values delete it, omitted keys are preserved. Node labels are commonly used as service placement constraints. Returns the updated node spec.",
-				),
-				mcplib.WithOutputSchema[nodeUpdateResult](),
-				mcplib.WithReadOnlyHintAnnotation(false),
-				mcplib.WithDestructiveHintAnnotation(false),
-				mcplib.WithIdempotentHintAnnotation(true),
-				mcplib.WithOpenWorldHintAnnotation(false),
-				mcplib.WithString("id",
-					mcplib.Required(),
-					mcplib.Description("Node ID or hostname."),
-				),
-				mcplib.WithObject(
-					"labels",
-					mcplib.Required(),
-					mcplib.Description(
-						"Merge-patch object mapping label keys to a string value (set) or null (delete).",
-					),
-				),
-			),
-			tier:    config.OpsConfiguration,
-			handler: s.toolUpdateNodeLabels,
-		},
-		{
-			tool: mcplib.NewTool(
-				"create_secret",
-				mcplib.WithToolTitle("Create a secret"),
-				mcplib.WithDescription(
-					"Create a Swarm secret. Swarm secrets cannot be changed once created, so rotating one is three calls in order: this tool for the replacement, update_service_secrets to repoint each service that uses the old one (describe the old secret first — its related array names them), then remove_secret once nothing references it. The value is write-only: no tool ever returns it, including this one. Pass encoding as base64 for binary content or anything JSON string escaping would mangle, such as a certificate or a key file; without it the value is stored exactly as given, since plenty of ordinary passwords are also valid base64 and guessing would silently store the wrong thing.",
-				),
-				mcplib.WithOutputSchema[createResult](),
-				mcplib.WithReadOnlyHintAnnotation(false),
-				mcplib.WithDestructiveHintAnnotation(false),
-				mcplib.WithIdempotentHintAnnotation(false),
-				mcplib.WithOpenWorldHintAnnotation(false),
-				mcplib.WithString(
-					"name",
-					mcplib.Required(),
-					mcplib.Description(
-						"Name for the new secret. Must not already exist in the cluster.",
-					),
-				),
-				mcplib.WithString("data",
-					mcplib.Required(),
-					mcplib.Description("The secret's value."),
-				),
-				mcplib.WithString(
-					"encoding",
-					mcplib.Enum("utf8", "base64"),
-					mcplib.Description(
-						"How `data` is encoded. Defaults to utf8, which stores it verbatim.",
-					),
-				),
-				mcplib.WithObject(
-					"labels",
-					mcplib.Description(
-						"Labels to set on the secret. Set com.docker.stack.namespace to attach it to a stack.",
-					),
-				),
-			),
-			tier:    config.OpsConfiguration,
-			handler: s.toolCreateSecret,
-		},
-		{
-			tool: mcplib.NewTool(
-				"create_config",
-				mcplib.WithToolTitle("Create a config"),
-				mcplib.WithDescription(
-					"Create a Swarm config. Like a secret, a config cannot be changed once created — replace it by creating a new one, repointing services with update_service_configs, and removing the old one. Unlike a secret, its content can be read back afterwards, so use a secret for anything sensitive. Pass encoding as base64 for a file whose content JSON string escaping would mangle.",
-				),
-				mcplib.WithOutputSchema[createResult](),
-				mcplib.WithReadOnlyHintAnnotation(false),
-				mcplib.WithDestructiveHintAnnotation(false),
-				mcplib.WithIdempotentHintAnnotation(false),
-				mcplib.WithOpenWorldHintAnnotation(false),
-				mcplib.WithString(
-					"name",
-					mcplib.Required(),
-					mcplib.Description(
-						"Name for the new config. Must not already exist in the cluster.",
-					),
-				),
-				mcplib.WithString("data",
-					mcplib.Required(),
-					mcplib.Description("The config's content."),
-				),
-				mcplib.WithString(
-					"encoding",
-					mcplib.Enum("utf8", "base64"),
-					mcplib.Description(
-						"How `data` is encoded. Defaults to utf8, which stores it verbatim.",
-					),
-				),
-				mcplib.WithObject(
-					"labels",
-					mcplib.Description(
-						"Labels to set on the config. Set com.docker.stack.namespace to attach it to a stack.",
-					),
-				),
-			),
-			tier:    config.OpsConfiguration,
-			handler: s.toolCreateConfig,
-		},
-		{
-			tool: mcplib.NewTool(
-				"update_service_secrets",
-				mcplib.WithToolTitle("Change which secrets a service receives"),
-				mcplib.WithDescription(
-					"Replace the complete set of secrets a service's containers receive. Swarm secrets cannot be changed in place, so rotating one is: create_secret for the replacement, this tool to repoint each service that uses the old one, then remove_secret once nothing references it. Describe the old secret first — its related array names the services to repoint. The list replaces rather than merges: pass every secret the service should end up with, because one you leave out is detached and the container loses it. Each entry names a secret rather than giving its ID. Triggers a rolling deploy, and you must be permitted to read a secret in order to attach it.",
-				),
-				mcplib.WithOutputSchema[serviceUpdateResult](),
-				mcplib.WithReadOnlyHintAnnotation(false),
-				mcplib.WithDestructiveHintAnnotation(true),
-				mcplib.WithIdempotentHintAnnotation(true),
-				mcplib.WithOpenWorldHintAnnotation(false),
-				mcplib.WithString("id",
-					mcplib.Required(),
-					mcplib.Description("Service ID or name."),
-				),
-				mcplib.WithArray("secrets",
-					mcplib.Required(),
-					mcplib.Description(
-						"The complete set of secrets the service should receive. Each entry takes `name` (the secret's name) and an optional `target` (the file it is mounted as, relative to /run/secrets unless absolute; defaults to the secret's name). An empty array detaches every secret.",
-					),
-					mcplib.Items(map[string]any{
-						"type": "object",
-						"properties": map[string]any{
-							"name":   map[string]any{"type": "string"},
-							"target": map[string]any{"type": "string"},
-						},
-						"required": []string{"name"},
-					}),
-				),
-			),
-			tier:    config.OpsConfiguration,
-			handler: s.toolUpdateServiceSecrets,
-		},
-		{
-			tool: mcplib.NewTool(
-				"update_service_configs",
-				mcplib.WithToolTitle("Change which configs a service receives"),
-				mcplib.WithDescription(
-					"Replace the complete set of configs a service's containers receive. Like secrets, Swarm configs cannot be changed in place: create the replacement with create_config, repoint the services here, then remove the old one. The list replaces rather than merges, so pass every config the service should end up with — one you leave out is detached. Each entry names a config rather than giving its ID. Triggers a rolling deploy, and you must be permitted to read a config in order to attach it.",
-				),
-				mcplib.WithOutputSchema[serviceUpdateResult](),
-				mcplib.WithReadOnlyHintAnnotation(false),
-				mcplib.WithDestructiveHintAnnotation(true),
-				mcplib.WithIdempotentHintAnnotation(true),
-				mcplib.WithOpenWorldHintAnnotation(false),
-				mcplib.WithString("id",
-					mcplib.Required(),
-					mcplib.Description("Service ID or name."),
-				),
-				mcplib.WithArray("configs",
-					mcplib.Required(),
-					mcplib.Description(
-						"The complete set of configs the service should receive. Each entry takes `name` (the config's name) and an optional `target` (the absolute path it is mounted at; defaults to the config's name). An empty array detaches every config.",
-					),
-					mcplib.Items(map[string]any{
-						"type": "object",
-						"properties": map[string]any{
-							"name":   map[string]any{"type": "string"},
-							"target": map[string]any{"type": "string"},
-						},
-						"required": []string{"name"},
-					}),
-				),
-			),
-			tier:    config.OpsConfiguration,
-			handler: s.toolUpdateServiceConfigs,
-		},
-		{
-			tool: mcplib.NewTool(
-				"update_service_mounts",
-				mcplib.WithToolTitle("Change a service's filesystem mounts"),
-				mcplib.WithDescription(
-					"Replace the complete set of filesystem mounts a service's containers receive: named volumes, host bind mounts and tmpfs. The list replaces rather than merges, so pass every mount the service should end up with — one you leave out is unmounted, and a container may lose data it was writing there. Triggers a rolling deploy. Note that a bind mount hands the container the host's filesystem at that path, and binding the Docker socket (/var/run/docker.sock) gives it control of the whole cluster; this tool will do it if asked, so check with the operator before mounting a host path they did not name.",
-				),
-				mcplib.WithOutputSchema[serviceUpdateResult](),
-				mcplib.WithReadOnlyHintAnnotation(false),
-				mcplib.WithDestructiveHintAnnotation(true),
-				mcplib.WithIdempotentHintAnnotation(true),
-				mcplib.WithOpenWorldHintAnnotation(false),
-				mcplib.WithString("id",
-					mcplib.Required(),
-					mcplib.Description("Service ID or name."),
-				),
-				mcplib.WithArray("mounts",
-					mcplib.Required(),
-					mcplib.Description(
-						"The complete set of mounts the service should receive. Each entry takes `type` (volume, bind or tmpfs), `target` (the absolute path inside the container), an optional `source` (the volume name for a volume, the host path for a bind; a tmpfs has none) and an optional `readOnly`. An empty array unmounts everything.",
-					),
-					mcplib.Items(map[string]any{
-						"type": "object",
-						"properties": map[string]any{
-							"type": map[string]any{
-								"type": "string",
-								"enum": []string{
-									string(mount.TypeVolume),
-									string(mount.TypeBind),
-									string(mount.TypeTmpfs),
-								},
-							},
-							"source":   map[string]any{"type": "string"},
-							"target":   map[string]any{"type": "string"},
-							"readOnly": map[string]any{"type": "boolean"},
-						},
-						"required": []string{"type", "target"},
-					}),
-				),
-			),
-			tier:    config.OpsConfiguration,
-			handler: s.toolUpdateServiceMounts,
-		},
-
-		// Tier 3 — Impactful.
-		{
-			tool: mcplib.NewTool(
-				"update_node",
-				mcplib.WithToolTitle("Update a node's role or availability"),
-				mcplib.WithDescription(
-					"Change a node's scheduling state, named by `section`. `availability` takes one of `active` (accepts new tasks), `pause` (rejects new tasks, keeps existing ones running) or `drain` (evicts every task to other nodes); draining a manager may destabilise the swarm, so prefer draining workers. `role` takes `worker` or `manager`: promoting adds a Raft voter, demoting removes one — never reduce the manager count below the quorum threshold, and for a three-manager cluster demote at most one at a time. Returns the node's role and availability as they now stand. Node labels are edited with update_node_labels, which needs a lower operations level than this.",
-				),
-				mcplib.WithOutputSchema[nodeUpdateResult](),
-				mcplib.WithReadOnlyHintAnnotation(false),
-				mcplib.WithDestructiveHintAnnotation(true),
-				mcplib.WithIdempotentHintAnnotation(true),
-				mcplib.WithOpenWorldHintAnnotation(false),
-				mcplib.WithString("id",
-					mcplib.Required(),
-					mcplib.Description("Node ID or hostname."),
-				),
-				mcplib.WithString("section",
-					mcplib.Required(),
-					mcplib.Enum(sectionAvailability, sectionRole),
-					mcplib.Description("Which part of the node's scheduling state to change."),
-				),
-				mcplib.WithString("value",
-					mcplib.Required(),
-					mcplib.Description(
-						"For `availability`: active, pause or drain. For `role`: worker or manager.",
-					),
-				),
-			),
-			tier:    config.OpsImpactful,
-			handler: s.toolUpdateNode,
-		},
-		{
-			tool: mcplib.NewTool("remove_service",
-				mcplib.WithToolTitle("Remove service"),
-				mcplib.WithOutputSchema[removalResult](),
-				mcplib.WithDescription(
-					"Delete a service and stop every task it owns. Does not delete the volumes, networks, configs, or secrets it referenced. Irreversible — no spec history is retained after removal.",
-				),
-				mcplib.WithReadOnlyHintAnnotation(false),
-				mcplib.WithDestructiveHintAnnotation(true),
-				mcplib.WithIdempotentHintAnnotation(false),
-				mcplib.WithOpenWorldHintAnnotation(false),
-				mcplib.WithString("id",
-					mcplib.Required(),
-					mcplib.Description("Service ID or name."),
-				),
-			),
-			tier: config.OpsImpactful,
-			handler: s.removeHandler("id", s.checkServiceWrite,
-				func(wc DockerWriteClient, ctx context.Context, id string) error {
-					return wc.RemoveService(ctx, id)
-				}),
-		},
-		{
-			tool: mcplib.NewTool("remove_config",
-				mcplib.WithToolTitle("Remove config"),
-				mcplib.WithOutputSchema[removalResult](),
-				mcplib.WithDescription(
-					"Delete a Docker config object. Fails if any service still references it. Irreversible — the config payload cannot be recovered.",
-				),
-				mcplib.WithReadOnlyHintAnnotation(false),
-				mcplib.WithDestructiveHintAnnotation(true),
-				mcplib.WithIdempotentHintAnnotation(false),
-				mcplib.WithOpenWorldHintAnnotation(false),
-				mcplib.WithString("id",
-					mcplib.Required(),
-					mcplib.Description("Config ID."),
-				),
-			),
-			tier: config.OpsImpactful,
-			handler: s.removeHandler("id", s.checkConfigWrite,
-				func(wc DockerWriteClient, ctx context.Context, id string) error {
-					return wc.RemoveConfig(ctx, id)
-				}),
-		},
-		{
-			tool: mcplib.NewTool("remove_secret",
-				mcplib.WithToolTitle("Remove secret"),
-				mcplib.WithOutputSchema[removalResult](),
-				mcplib.WithDescription(
-					"Delete a Docker secret. Fails if any service still references it. Irreversible — the secret payload cannot be recovered.",
-				),
-				mcplib.WithReadOnlyHintAnnotation(false),
-				mcplib.WithDestructiveHintAnnotation(true),
-				mcplib.WithIdempotentHintAnnotation(false),
-				mcplib.WithOpenWorldHintAnnotation(false),
-				mcplib.WithString("id",
-					mcplib.Required(),
-					mcplib.Description("Secret ID."),
-				),
-			),
-			tier: config.OpsImpactful,
-			handler: s.removeHandler("id", s.checkSecretWrite,
-				func(wc DockerWriteClient, ctx context.Context, id string) error {
-					return wc.RemoveSecret(ctx, id)
-				}),
-		},
-		{
-			tool: mcplib.NewTool("remove_network",
-				mcplib.WithToolTitle("Remove network"),
-				mcplib.WithOutputSchema[removalResult](),
-				mcplib.WithDescription(
-					"Delete an overlay network. Fails if any service is still attached. Built-in networks (`ingress`, `docker_gwbridge`) cannot be removed.",
-				),
-				mcplib.WithReadOnlyHintAnnotation(false),
-				mcplib.WithDestructiveHintAnnotation(true),
-				mcplib.WithIdempotentHintAnnotation(false),
-				mcplib.WithOpenWorldHintAnnotation(false),
-				mcplib.WithString("id",
-					mcplib.Required(),
-					mcplib.Description("Network ID."),
-				),
-			),
-			tier: config.OpsImpactful,
-			handler: s.removeHandler("id", s.checkNetworkWrite,
-				func(wc DockerWriteClient, ctx context.Context, id string) error {
-					return wc.RemoveNetwork(ctx, id)
-				}),
-		},
-		{
-			tool: mcplib.NewTool("remove_volume",
-				mcplib.WithToolTitle("Remove volume"),
-				mcplib.WithOutputSchema[removalResult](),
-				mcplib.WithDescription(
-					"Delete a Docker volume by name. Fails if any container is currently using it unless `force` is set. Volume contents are irrecoverable.",
-				),
-				mcplib.WithReadOnlyHintAnnotation(false),
-				mcplib.WithDestructiveHintAnnotation(true),
-				mcplib.WithIdempotentHintAnnotation(false),
-				mcplib.WithOpenWorldHintAnnotation(false),
-				mcplib.WithString("name",
-					mcplib.Required(),
-					mcplib.Description("Volume name."),
-				),
-				mcplib.WithBoolean(
-					"force",
-					mcplib.Description(
-						"Force removal even if the volume is in use. Default false.",
-					),
-				),
-			),
-			tier:    config.OpsImpactful,
-			handler: s.toolRemoveVolume,
-		},
-	}
+	return slices.Concat(
+		s.readTools(),
+		s.operationalTools(),
+		s.configurationTools(),
+		s.impactfulTools(),
+	)
 }
 
 // requireWriteClient returns the write client or an error explaining that
@@ -1164,128 +295,6 @@ func (s *Server) requireWriteClient() (DockerWriteClient, error) {
 		return nil, fmt.Errorf("MCP server has no write client configured")
 	}
 	return s.writeClient, nil
-}
-
-// checkWrite enforces the "write" permission on resourceType:resourceName for
-// the identity in ctx. Returns nil if ACL is disabled or no identity is on the
-// context (the bearer middleware would have rejected the request earlier in
-// that case). All MCP tool handlers route through this helper so denial errors
-// have a uniform shape.
-func (s *Server) checkWrite(ctx context.Context, resourceType, resourceName string) error {
-	if s.acl == nil {
-		return nil
-	}
-	identity := auth.IdentityFromContext(ctx)
-	if identity == nil {
-		return nil
-	}
-	if !s.acl.Can(identity, "write", resourceType+":"+resourceName) {
-		return fmt.Errorf("write access denied for %s:%s", resourceType, resourceName)
-	}
-	return nil
-}
-
-// checkServiceWrite resolves the service name from the cache so the ACL key
-// is `service:<name>` rather than `service:<id>`, matching REST behaviour.
-func (s *Server) checkServiceWrite(ctx context.Context, id string) error {
-	name := id
-	if svc, ok := s.cache.GetService(id); ok && svc.Spec.Name != "" {
-		name = svc.Spec.Name
-	}
-	return s.checkWrite(ctx, "service", name)
-}
-
-// checkNodeWrite resolves the node hostname from the cache so the ACL key is
-// `node:<hostname>` rather than `node:<id>`, matching REST behaviour.
-func (s *Server) checkNodeWrite(ctx context.Context, id string) error {
-	name := id
-	if node, ok := s.cache.GetNode(id); ok && node.Description.Hostname != "" {
-		name = node.Description.Hostname
-	}
-	return s.checkWrite(ctx, "node", name)
-}
-
-// checkConfigWrite resolves the config Spec.Name from the cache so the ACL key
-// is `config:<name>` rather than `config:<id>`, matching REST behaviour.
-func (s *Server) checkConfigWrite(ctx context.Context, id string) error {
-	name := id
-	if cfg, ok := s.cache.GetConfig(id); ok && cfg.Spec.Name != "" {
-		name = cfg.Spec.Name
-	}
-	return s.checkWrite(ctx, "config", name)
-}
-
-// checkSecretWrite resolves the secret Spec.Name from the cache so the ACL key
-// is `secret:<name>` rather than `secret:<id>`, matching REST behaviour.
-func (s *Server) checkSecretWrite(ctx context.Context, id string) error {
-	name := id
-	if sec, ok := s.cache.GetSecret(id); ok && sec.Spec.Name != "" {
-		name = sec.Spec.Name
-	}
-	return s.checkWrite(ctx, "secret", name)
-}
-
-// checkNetworkWrite resolves the network Name from the cache so the ACL key is
-// `network:<name>` rather than `network:<id>`, matching REST behaviour.
-func (s *Server) checkNetworkWrite(ctx context.Context, id string) error {
-	name := id
-	if net, ok := s.cache.GetNetwork(id); ok && net.Name != "" {
-		name = net.Name
-	}
-	return s.checkWrite(ctx, "network", name)
-}
-
-// checkTaskWrite routes the ACL check to the task's parent service so an MCP
-// `remove_task` evaluates the same key REST DELETE /tasks/{id} does
-// (`service:<name>`). Falls back to `task:<id>` when the cache cannot resolve
-// the parent service — that path is also what REST takes as a last resort.
-func (s *Server) checkTaskWrite(ctx context.Context, id string) error {
-	task, ok := s.cache.GetTask(id)
-	if !ok {
-		return s.checkWrite(ctx, "task", id)
-	}
-	svc, ok := s.cache.GetService(task.ServiceID)
-	if !ok || svc.Spec.Name == "" {
-		return s.checkWrite(ctx, "task", id)
-	}
-	return s.checkWrite(ctx, "service", svc.Spec.Name)
-}
-
-// checkServiceRead resolves the service Spec.Name from the cache so the ACL
-// key is `service:<name>` rather than `service:<id>`, matching the resource
-// read path in lookupResource. Used by toolGetLogs which doesn't go through
-// lookupResource.
-//
-// It resolves rather than looking the ID up, because get_logs advertises its
-// `service` argument as "Service ID or name" and Docker honours both — so a
-// plain GetService turned a name, the identifier find and the completions
-// hand back, into "service not found" on every cluster that has an ACL policy
-// and into a working call on every cluster that has not.
-func (s *Server) checkServiceRead(ctx context.Context, id string) error {
-	if s.acl == nil {
-		return nil
-	}
-	svc, ok, err := s.cache.ResolveService(id)
-	if err != nil {
-		return err
-	}
-	if !ok {
-		return fmt.Errorf("service %q not found", id)
-	}
-	return s.checkRead(ctx, "service", svc.Spec.Name)
-}
-
-// checkTaskRead enforces the read permission on a task, keyed as `task:<id>`.
-//
-// It does not walk to the parent service the way checkTaskWrite does, because
-// the evaluator already does: acl.grantMatchesResource resolves a task through
-// its parent service and that service's stack, so the task's own key is
-// strictly the broader check. It is also the key every other task read passes
-// — REST's HandleTaskLogs and the cetacean://tasks/{id} resource both do — and
-// walking to the parent here would leave a `task:*` grant able to read task
-// logs on every path except this tool.
-func (s *Server) checkTaskRead(ctx context.Context, id string) error {
-	return s.checkRead(ctx, "task", id)
 }
 
 // --- tool handlers ---
